# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Vexa-VM API",
    "version": "1.0.0",
    "summary": "Single-node KVM administration and customer self-service API",
    "description": "Administrator automation uses /api/v1 with a scoped bearer API key or administrator session. Customer self-service uses /api/public with a VM-scoped status session. Long-running KVM operations return durable jobs."
  },
  "servers": [
    {
      "url": "/",
      "description": "The current Vexa-VM HTTPS origin"
    }
  ],
...
  "security": [
    { "adminApiKey": [] },
    { "adminSession": [] }
  ],
  "paths": {
...
    "/api/v1/auth/login": {
      "post": {
        "tags": ["Authentication"],
        "operationId": "loginAdmin",
        "summary": "Create an administrator browser session",
        "security": [],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/LoginRequest" } } }
        },
        "responses": {
          "200": {
            "description": "Administrator and readable CSRF cookies issued",
            "headers": {
              "Set-Cookie": { "schema": { "type": "string" }, "description": "HttpOnly administrator session plus readable CSRF cookie; Secure is set when secure cookies are enabled" },
              "Cache-Control": { "schema": { "type": "string", "const": "no-store" } }
            },
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/LoginResponse" } } }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "429": { "$ref": "#/components/responses/RateLimited" }
        }
      }
    },
    "/api/v1/auth/me": {
      "get": {
        "tags": ["Authentication"],
        "operationId": "getCurrentAdmin",
        "summary": "Read the current authenticated administrator/API-key actor",
        "responses": {
          "200": { "description": "Current authenticated actor", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/AuthContext" } } } },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      }
    },
    "/api/v1/auth/logout": {
      "post": {
        "tags": ["Authentication"],
        "operationId": "logoutAdmin",
        "summary": "Revoke the current administrator session",
        "description": "A browser-session call revokes that session. A bearer API-key call has no session to revoke, but still returns success and clears session cookies in the response.",
        "parameters": [{ "$ref": "#/components/parameters/CsrfToken" }],
        "responses": {
          "200": { "description": "Session revoked when the credential was a browser session", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Success" } } } },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" }
        }
      }
    },
    "/api/v1/host": {
      "get": {
        "tags": ["Host"],
        "operationId": "getHost",
        "summary": "Read detected host resources and network inventory",
        "x-required-scopes": ["host:read"],
        "responses": {
          "200": { "description": "Host inventory", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/HostEnvelope" } } } },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "503": { "$ref": "#/components/responses/Unavailable" }
        }
      }
    },
...
    "/api/v1/vms": {
      "get": {
        "tags": ["Virtual machines"],
        "operationId": "listVms",
        "summary": "List virtual machines",
        "x-required-scopes": ["vms:read"],
        "responses": {
          "200": { "description": "Complete VM inventory; page.next_cursor is always null", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/VmList" } } } },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" }
        }
      },
      "post": {
        "tags": ["Virtual machines"],
        "operationId": "createVm",
        "summary": "Queue virtual-machine creation",
        "description": "For blank, cloud-init, and automatic images, omitting password generates a stored provisioning password. Manual installer images reject a password value and neither generate nor store one. Capacity admission is serialized through provisional VM and job publication; non-error CPU/RAM allocations and unmaterialized creating-disk reservations are counted before a request is accepted.",
        "x-required-scopes": ["vms:write"],
        "parameters": [
          { "$ref": "#/components/parameters/IdempotencyKey" },
          { "$ref": "#/components/parameters/CsrfToken" }
        ],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/VmCreate" } } } },
        "responses": {
          "202": { "description": "VM record created and provisioning operation queued", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/VmCreateAccepted" } } } },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "409": { "$ref": "#/components/responses/Conflict" }
        }
      }
    },
    "/api/v1/vms/{vm_id}": {
      "parameters": [{ "$ref": "#/components/parameters/VmId" }],
      "get": {
        "tags": ["Virtual machines"],
        "operationId": "getVm",
        "summary": "Read one virtual machine",
        "x-required-scopes": ["vms:read"],
        "responses": {
          "200": { "description": "VM detail", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/VmEnvelope" } } } },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      },
      "patch": {
        "tags": ["Virtual machines"],
        "operationId": "updateVm",
        "summary": "Update VM metadata or resources",
        "description": "A hostname change is persisted and attempted live through authenticated Vexa Guest Tools. Inspect guest_tools rather than inferring live guest success from HTTP 200. Resource changes are represented by the optional operation job.",
        "x-required-scopes": ["vms:write"],
        "parameters": [
          { "$ref": "#/components/parameters/IdempotencyKey" },
          { "$ref": "#/components/parameters/CsrfToken" }
        ],
        "requestBody": { "required": true, "content": { "application/merge-patch+json": { "schema": { "$ref": "#/components/schemas/VmPatch" } }, "application/json": { "schema": { "$ref": "#/components/schemas/VmPatch" } } } },
        "responses": {
          "200": { "description": "VM updated; operation is a resize job or null", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/VmUpdateResult" } } } },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "409": { "$ref": "#/components/responses/Conflict" }
        }
      },
      "delete": {
        "tags": ["Virtual machines"],
        "operationId": "deleteVm",
        "summary": "Queue virtual-machine deletion",
        "description": "A retry with the same Idempotency-Key returns the original deletion job even after the VM row has been removed.",
        "x-required-scopes": ["vms:write"],
        "parameters": [
          { "$ref": "#/components/parameters/IdempotencyKey" },
          { "$ref": "#/components/parameters/CsrfToken" }
        ],
        "responses": {
          "202": { "$ref": "#/components/responses/AcceptedOperation" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "409": { "$ref": "#/components/responses/Conflict" }
        }
      }
    },
    "/api/v1/vms/{vm_id}/actions/{action}": {
      "post": {
        "tags": ["Virtual machines"],
        "operationId": "powerVm",
        "summary": "Apply a VM power action",
        "description": "suspend and pause freeze a running guest and set inventory state to paused; resume continues it. Maintenance is configured separately and does not change guest power state.",
        "x-required-scopes": ["vms:power"],
        "parameters": [
          { "$ref": "#/components/parameters/VmId" },
          { "$ref": "#/components/parameters/PowerAction" },
          { "$ref": "#/components/parameters/IdempotencyKey" },
          { "$ref": "#/components/parameters/CsrfToken" }
        ],
        "responses": {
          "202": { "$ref": "#/components/responses/AcceptedOperation" },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "409": { "$ref": "#/components/responses/Conflict" }
        }
      }
    },
    "/api/v1/vms/{vm_id}/maintenance": {
      "put": {
        "tags": ["Virtual machines"],
        "operationId": "setVmMaintenance",
        "summary": "Enable or clear the customer-mutation maintenance window",
        "description": "Maintenance does not pause or power off the guest. Customer reads remain available, but customer mutations return 409 until maintenance is cleared.",
        "x-required-scopes": ["vms:write"],
        "parameters": [{ "$ref": "#/components/parameters/VmId" }, { "$ref": "#/components/parameters/CsrfToken" }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/VmMaintenanceUpdate" } } } },
        "responses": {
          "200": { "description": "Maintenance policy saved", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/VmEnvelope" } } } },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
    "/api/v1/vms/{vm_id}/disk-protection": {
      "put": {
        "tags": ["Virtual machines"],
        "operationId": "setVmDiskProtection",
        "summary": "Set deletion and reinstall safeguards",
        "description": "Both safeguards default to false. A deletion lock rejects VM deletion; snapshot_before_reinstall aborts disk replacement unless its pre-reinstall snapshot succeeds.",
        "x-required-scopes": ["vms:write"],
        "parameters": [{ "$ref": "#/components/parameters/VmId" }, { "$ref": "#/components/parameters/CsrfToken" }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/VmDiskProtectionUpdate" } } } },
        "responses": {
          "200": { "description": "Disk-protection policy saved", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/VmEnvelope" } } } },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
    "/api/v1/vms/{vm_id}/network-security": {
      "parameters": [{ "$ref": "#/components/parameters/VmId" }],
      "get": {
        "tags": ["Virtual machines", "Network"],
        "operationId": "getVmNetworkSecurity",
        "summary": "Read a VM's firewall and DDoS profile plus rules",
        "x-required-scopes": ["vms:read"],
        "responses": {
          "200": { "description": "Desired/applied profile and rules", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/VmNetworkSecurityEnvelope" } } } },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      },
      "patch": {
        "tags": ["Virtual machines", "Network"],
        "operationId": "updateVmNetworkSecurity",
        "summary": "Update opt-in VM firewall and DDoS policy",
        "description": "All protection defaults are off. Enabling DDoS requires at least one supported limiter/protection. A non-null concurrent_connection_limit is rejected in this release.",
        "x-required-scopes": ["vms:write"],
        "parameters": [{ "$ref": "#/components/parameters/CsrfToken" }],
        "requestBody": { "required": true, "content": { "application/merge-patch+json": { "schema": { "$ref": "#/components/schemas/VmNetworkSecurityPatch" } }, "application/json": { "schema": { "$ref": "#/components/schemas/VmNetworkSecurityPatch" } } } },
        "responses": {
          "200": { "description": "Desired profile and host reconciliation summary", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/VmNetworkSecurityUpdateResult" } } } },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
    "/api/v1/vms/{vm_id}/firewall/rules": {
      "parameters": [{ "$ref": "#/components/parameters/VmId" }],
      "get": {
        "tags": ["Virtual machines", "Network"],
        "operationId": "listVmFirewallRules",
        "summary": "List VM firewall rules",
        "x-required-scopes": ["vms:read"],
        "responses": {
          "200": { "description": "Rules", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/FirewallRuleList" } } } },
          "401": { "$ref": "#/components/responses/Unauthorized" }, "403": { "$ref": "#/components/responses/Forbidden" }, "404": { "$ref": "#/components/responses/NotFound" }
        }
      },
      "post": {
        "tags": ["Virtual machines", "Network"],
        "operationId": "createVmFirewallRule",
        "summary": "Create a disabled-by-default VM firewall rule",
        "x-required-scopes": ["vms:write"],
        "parameters": [{ "$ref": "#/components/parameters/CsrfToken" }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/FirewallRuleCreate" } } } },
        "responses": {
          "201": { "description": "Rule created and policy reconciled", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/FirewallRuleMutationResult" } } } },
          "400": { "$ref": "#/components/responses/BadRequest" }, "401": { "$ref": "#/components/responses/Unauthorized" }, "403": { "$ref": "#/components/responses/Forbidden" }, "404": { "$ref": "#/components/responses/NotFound" }, "409": { "$ref": "#/components/responses/Conflict" }
        }
      }
    },
...
    "/api/v1/vms/{vm_id}/traffic/reset": {
      "post": {
        "tags": ["Virtual machines"],
        "operationId": "resetVmTraffic",
        "summary": "Reset VM traffic accounting and restore a quota-blocked link",
        "x-required-scopes": ["vms:write"],
        "parameters": [
          { "$ref": "#/components/parameters/VmId" },
          { "$ref": "#/components/parameters/CsrfToken" }
        ],
        "responses": {
          "200": { "description": "Traffic reset and current quota enforcement status" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "503": { "description": "The counter was reset but libvirt could not restore the network link; reconciliation will retry" }
        }
      }
    },
...
    "/api/v1/vms/{vm_id}/guest-tools": {
      "parameters": [{ "$ref": "#/components/parameters/VmId" }],
      "get": {
        "tags": ["Virtual machines"], "operationId": "getVmGuestTools", "summary": "Read Vexa Guest Tools state", "x-required-scopes": ["vms:read"],
        "responses": { "200": { "description": "Guest Tools state; the channel secret is never returned", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/GuestToolsStatusEnvelope" } } } }, "401": { "$ref": "#/components/responses/Unauthorized" }, "403": { "$ref": "#/components/responses/Forbidden" }, "404": { "$ref": "#/components/responses/NotFound" } }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "adminApiKey": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "Vexa-VM API key",
        "description": "A scoped administrator API key. The plaintext value is returned only when the key is created. An optional ip_allowlist is checked against the socket peer, or a forwarded address only when the immediate peer is loopback."
      },
      "adminSession": {
        "type": "apiKey",
        "in": "cookie",
        "name": "vexa_session",
        "description": "Secure HttpOnly administrator browser session. Writes also require X-CSRF-Token."
      },
      "customerSession": {
        "type": "apiKey",
        "in": "cookie",
        "name": "vexa_status",
        "description": "Secure HttpOnly VM-scoped customer session. Writes also require X-CSRF-Token."
      }
    },
    "parameters": {
...
      "VmId": { "name": "vm_id", "in": "path", "required": true, "description": "VM UUID or case-insensitive VM name.", "schema": { "type": "string" } },
...
    },
    "responses": {
...
      "BadRequest": { "description": "Invalid request", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorEnvelope" } } } },
...
    },
    "schemas": {
...
      "LoginRequest": {
        "type": "object",
        "required": ["username", "password"],
        "properties": {
          "username": { "type": "string", "minLength": 1, "maxLength": 80 },
          "password": { "type": "string", "minLength": 1, "maxLength": 4096, "writeOnly": true }
        },
        "additionalProperties": false
      },
...
    }
  }
}